refactor(multilayer-map): rename feature folder
