+++ lib/features/digital_guide/data/repository/digital_guide_repository.dart
@@ -2,7 +2,7 @@ import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
 import "../../../../api_base_rest/client/json.dart";
-import "../../../buildings_view/data/repositories/buildings_repository.dart";
+import "../../../multilayer_map/data/repositories/buildings_repository.dart";
 import "../api/digital_guide_get_and_cache.dart";
 import "../models/digital_guide_response.dart";
