+++ lib/features/digital_guide/presentation/digital_guide_view.dart
@@ -12,7 +12,7 @@ import "../../../widgets/detail_views/detail_view_app_bar.dart";
 import "../../../widgets/horizontal_symmetric_safe_area.dart";
 import "../../../widgets/my_error_widget.dart";
 import "../../../widgets/zoomable_images.dart";
-import "../../buildings_view/data/model/building.dart";
+import "../../multilayer_map/data/model/building.dart";
 import "../data/models/digital_guide_response.dart";
 import "../data/repository/digital_guide_repository.dart";
 import "widgets/accessibility_button.dart";
