+++ lib/features/digital_guide/tabs/localization/presentation/localization_expansion_tile_content.dart
@@ -5,8 +5,8 @@ import "package:latlong2/latlong.dart";
 import "../../../../../config/ui_config.dart";
 import "../../../../../theme/app_theme.dart";
 import "../../../../../utils/context_extensions.dart";
-import "../../../../buildings_view/data/model/building.dart";
 import "../../../../map_view/utils/google_maps_link_utils.dart";
+import "../../../../multilayer_map/data/model/building.dart";
 import "../../../data/models/digital_guide_response.dart";
 import "../../../presentation/widgets/digital_guide_image.dart";
