+++ lib/features/digital_guide_objects/presentation/digital_guide_object_view.dart
@@ -12,13 +12,13 @@ import "../../../widgets/detail_views/detail_view_app_bar.dart";
 import "../../../widgets/horizontal_symmetric_safe_area.dart";
 import "../../../widgets/my_error_widget.dart";
 import "../../../widgets/zoomable_images.dart";
-import "../../buildings_view/data/model/building.dart";
 import "../../digital_guide/presentation/widgets/accessibility_button.dart";
 import "../../digital_guide/presentation/widgets/accessibility_information_cards_list.dart";
 import "../../digital_guide/presentation/widgets/digital_guide_data_source_link.dart";
 import "../../digital_guide/presentation/widgets/digital_guide_loading_view.dart";
 import "../../digital_guide/presentation/widgets/headlines_section.dart";
 import "../../digital_guide/presentation/widgets/report_change_button.dart";
+import "../../multilayer_map/data/model/building.dart";
 import "../data/models/digital_guide_object_model.dart";
 import "../data/repositories/digital_guide_object_repository.dart";
 import "digital_g_objects_featers_list.dart";
