+++ lib/features/home_view/widgets/buildings_section/buildings_section.dart
@@ -5,8 +5,8 @@ import "package:flutter_riverpod/flutter_riverpod.dart";
 import "../../../../utils/context_extensions.dart";
 import "../../../../widgets/my_error_widget.dart";
 import "../../../../widgets/subsection_header.dart";
-import "../../../buildings_view/data/model/building.dart";
-import "../../../buildings_view/data/repositories/buildings_repository.dart";
+import "../../../multilayer_map/data/model/building.dart";
+import "../../../multilayer_map/data/repositories/buildings_repository.dart";
 import "../../../navigator/utils/navigation_commands.dart";
 import "../loading_widgets/scrollable_section_loading.dart";
 import "../paddings.dart";
