+++ lib/features/map_view/map_view.dart
@@ -7,8 +7,8 @@ import "package:flutter_riverpod/flutter_riverpod.dart";
 import "../../config/map_view_config.dart";
 import "../../theme/app_theme.dart";
 import "../bottom_scroll_sheet/bottom_scroll_sheet.dart";
-import "../bottom_scroll_sheet/map_view_pop_behaviour.dart";
 import "../bottom_scroll_sheet/map_data_sheet_list.dart";
+import "../bottom_scroll_sheet/map_view_pop_behaviour.dart";
 import "controllers/bottom_sheet_controller.dart";
 import "controllers/controllers_set.dart";
 import "widgets/map_config.dart";
