+++ lib/features/navigator/utils/navigation_commands.dart
@@ -6,7 +6,6 @@ import "package:logger/logger.dart";
 import "../../../utils/launch_url_util.dart";
 import "../../analytics/data/umami.dart";
 import "../../analytics/data/umami_events.dart";
-import "../../buildings_view/data/model/building.dart";
 import "../../digital_guide/data/models/level.dart" as digital_guide;
 import "../../digital_guide/data/models/level_with_regions.dart";
 import "../../digital_guide/data/models/region.dart";
@@ -24,6 +23,7 @@ import "../../digital_guide/tabs/structure/data/models/ramp.dart";
 import "../../digital_guide/tabs/structure/data/models/stairway.dart";
 import "../../digital_guide/tabs/structure/data/models/toilet.dart";
 import "../../digital_guide/tabs/transportation/data/models/digital_guide_transportation.dart";
+import "../../multilayer_map/data/model/building.dart";
 import "../../parkings/parkings_view/models/parking.dart";
 import "../../science_club/science_clubs_view/model/science_clubs.dart" show ScienceClub;
 import "../../science_club/science_clubs_view/repository/science_clubs_repository.dart";
